Separation between the camera component and data provider
