+++ custom_components/xiaomi_cloud_map_extractor/const.py
@@ -212,3 +212,23 @@ API_EXCEPTIONS = {
 }
 
 MM = 50
+
+DEFAULT_TRIMS = {
+    CONF_LEFT: 0,
+    CONF_RIGHT: 0,
+    CONF_TOP: 0,
+    CONF_BOTTOM: 0
+}
+
+DEFAULT_SIZES = {
+    CONF_SIZE_VACUUM_RADIUS: 6,
+    CONF_SIZE_PATH_WIDTH: 1,
+    CONF_SIZE_IGNORED_OBSTACLE_RADIUS: 3,
+    CONF_SIZE_IGNORED_OBSTACLE_WITH_PHOTO_RADIUS: 3,
+    CONF_SIZE_OBSTACLE_RADIUS: 3,
+    CONF_SIZE_OBSTACLE_WITH_PHOTO_RADIUS: 3,
+    CONF_SIZE_CHARGER_RADIUS: 6
+}
+
+DEFAULT_MAP_TRANSFORM = {CONF_SCALE: 1, CONF_ROTATE: 0, CONF_TRIM: DEFAULT_TRIMS}
+DEFAULT_STORE_MAP_PATH = "/tmp"
